fix(api): allow login/logout from the frontend to propagate to the backend (#10)
* fix(auth): have User API handle session data

This is a workaround to have logins and logouts on the frontend "propagate" to
the backend, so that admins only need to log in or out once for both.

* chore(dependencies): make social auth libraries explicit dependencies

This patch adds "djangorestframework_simplejwt" and
"social-auth-app-django" as explicit dependencies.

These libraries have been extensively used in the application, and not
adding them to the pyproject.toml file is lying to contributors about
which dependencies are in use.

* Fixed login/logut

* API login returning CSRF_COOKIE

* logout API in post

---------

Co-authored-by: GigiusB <g.bronzini@gmail.com>