feat(core): replace work and overtime hours with day/night shift hours (#13)
Overtime only comes in when generating weekly reports, and it's not the
user's responsibility to track it, so we have to get rid of it.

Night shift, however, is a pretty important feature which KRM3 was
missing at the start, as it usually is paid more than normal work.

This has led to the changes to time entry fields in this patch:
* `overtime_hours` is now called `night_shift_hours`;
* `work_hours` are now called `day_shift_hours`: the new name better
  reflects the distinction with night shift hours, and it makes it very
  clear that daytime and nighttime work hours are handled separately.

Unlike the old KRM, which treats night shift hours as part of the
declared work hours, KRM3 considers the two as separate, which feels
more intuitive and does not require changing the current logic to
compute work hour totals.