feat(api): Add a restricted active resources list endpoint (#27)
* fix(core): adjust permissions to view/manage timesheets

A contentious point is whether to allow non-project viewers or managers
to view or manage timesheets. We are currently solving this by denying
them access, as they are not allowed to see any tasks due to ACL
filtering.

* feat(api): add a specific endpoint for active resources

It may be overkill - we could have added a query param instead - but
this specific endpoint only works for users who can at least view
both projects and timesheets. Adding a normal query filter would prevent
normal searches for anyone, and we have no clear requirements about
that.

* fix(api): restrict access to time entry creation and deletion

If you don't own a timesheet, you can add or delete time entries on it
only if you have:
- at least read only access to all projects;
- read/write access to all timesheets.

* fix(api): send permission strings along with user data

This affects the "me" endpoint.